+++ src/main/java/dev/muon/medievalorigins/mixin/compat/icarus/IcarusClientMixin.java
@@ -10,13 +10,14 @@ import dev.muon.medievalorigins.power.IcarusWingsPowerType;
 import dev.muon.medievalorigins.power.PixieWingsPowerType;
 import dev.muon.medievalorigins.util.ItemDataUtil;
 import io.github.apace100.apoli.component.PowerHolderComponent;
-import io.github.apace100.apoli.power.type.PowerType;
 import net.minecraft.client.player.AbstractClientPlayer;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ArmorItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 import org.spongepowered.asm.mixin.Mixin;
+
+import java.util.List;
 import org.spongepowered.asm.mixin.injection.At;
 
 @Mixin(IcarusClient.class)
@@ -42,13 +43,13 @@ public abstract class IcarusClientMixin {
     @ModifyReturnValue(method = "getWingsForRendering", at = @At(value = "RETURN"))
     private static ItemStack renderOriginWings(ItemStack original, LivingEntity entity) {
         if (original.isEmpty()) {
-            var icarusWings = PowerHolderComponent.getPowerTypes(entity, IcarusWingsPowerType.class).stream().filter(PowerType::isActive).findFirst();
-            if (icarusWings.isPresent()) {
-                return icarusWings.get().getWingsType();
+            List<IcarusWingsPowerType> icarusWings = PowerHolderComponent.getPowerTypes(entity, IcarusWingsPowerType.class);
+            if (!icarusWings.isEmpty()) {
+                return icarusWings.getFirst().getWingsType();
             }
         } else {
-            boolean hasPixieWings = PowerHolderComponent.hasPowerType(entity, PixieWingsPowerType.class, PowerType::isActive);
-            boolean hasFaeWings = PowerHolderComponent.hasPowerType(entity, FaeWingsPowerType.class, PowerType::isActive);
+            boolean hasPixieWings = PowerHolderComponent.hasPowerType(entity, PixieWingsPowerType.class);
+            boolean hasFaeWings = PowerHolderComponent.hasPowerType(entity, FaeWingsPowerType.class);
 
             if (hasPixieWings || hasFaeWings) {
                 return new ItemStack(Items.AIR);